feat: implement createKiwi function for kiwi creation and update StepThree to utilize it
