+++ apps/web/src/routes/my-kiwis/-modals/create-kiwi/steps/three.tsx
@@ -7,27 +7,24 @@ import { useAtomValue, useSetAtom } from "@bookiwi/jotai";
 import { createKiwiAtom, setShareCodeAtom, stepAtom } from "../atoms";
 import { Step } from "../types";
 
-import tempUser from "#/DB/users";
 import { Button } from "#/components/ui/button";
 import { DialogFooter } from "#/components/ui/dialog";
-import { color, primaryColor } from "#/constants/color";
-import idb, { IDBStore } from "#/managers/idb";
-import { EpubIDBData, KiwiIDBData, ParticipantIDBData } from "#/types/idb";
-import { fileToBookInfo } from "#/utils/epubjs";
+import { primaryColor } from "#/constants/color";
+import { createKiwi } from "#/routes/my-kiwis/-apis/create-kiwi";
 
 function StepThree() {
   const newKiwi = useAtomValue(createKiwiAtom);
   const setShareCode = useSetAtom(setShareCodeAtom);
   const setStep = useSetAtom(stepAtom);
-  const abortControllerRef = useRef<AbortController | null>(null);
+  // const abortControllerRef = useRef<AbortController | null>(null);
   const hasExecutedRef = useRef(false);
   const router = useRouter();
   const [isFailed, setIsFailed] = useState(false);
 
-  const handleCancel = () => {
-    abortControllerRef.current?.abort();
-    setStep(Step.Two);
-  };
+  // const handleCancel = () => {
+  //   abortControllerRef.current?.abort();
+  //   setStep(Step.Two);
+  // };
 
   useEffect(() => {
     // Strict Mode에서 중복 실행 방지
@@ -37,8 +34,8 @@ function StepThree() {
     hasExecutedRef.current = true;
 
     const handleSubmit = async () => {
-      const controller = new AbortController();
-      abortControllerRef.current = controller;
+      // const controller = new AbortController();
+      // abortControllerRef.current = controller;
       try {
         // 여기서 실제로 API 호출 등의 작업을 수행
 
@@ -48,84 +45,12 @@ function StepThree() {
         //   signal: controller.signal, // ✅ 여기가 중요
         // });
         // const book = new Book(state.selectedFile);
-
-        const bookInfo = await fileToBookInfo(newKiwi.file!);
-
-        // 공유 코드 생성 (실제로는 API에서 받아와야 함)
-        const generatedShareCode = `KIWI-${Math.random().toString(36).substring(2, 8).toUpperCase()}`;
-        const generatedKiwiId = Math.random()
-          .toString(36)
-          .substring(2, 8)
-          .toUpperCase();
-        const generatedEpubId = Math.random()
-          .toString(36)
-          .substring(2, 8)
-          .toUpperCase();
-
-        const generatedParticipantId = Math.random()
-          .toString(36)
-          .substring(2, 8)
-          .toUpperCase();
-
-        if (abortControllerRef.current?.signal.aborted) {
-          return;
-        }
-        const kiwiIDBData: KiwiIDBData = {
-          id: generatedKiwiId,
-          name: newKiwi.kiwiName,
-          description: newKiwi.kiwiDescription,
-          maxParticipants: newKiwi.maxParticipants,
-          detailDescription: newKiwi.kiwiDetailDescription,
-          password: newKiwi.password,
-          shareCode: generatedShareCode,
-          createdAt: new Date().toISOString(),
-          coverImage: bookInfo.coverImageBlob,
-          bookMetadata: {
-            title: bookInfo.title,
-            author: bookInfo.author,
-            publisher: bookInfo.publisher,
-            toc: bookInfo.toc,
-          },
-          epubId: generatedEpubId,
-          adminId: tempUser.id,
-          participantIds: [generatedParticipantId],
-        };
-
-        const participantIDBData: ParticipantIDBData = {
-          id: generatedParticipantId,
-          kiwiId: generatedKiwiId,
-          userId: tempUser.id,
-          name: tempUser.name,
-          profileImage: tempUser.profileImage,
-          color: color[0],
-          record: {
-            currentCfi: null,
-            percentage: null,
-            bookmarks: [],
-          },
-          settings: {
-            isSinglePage: false,
-            fontFamily: null,
-            fontSize: null,
-            lineHeight: null,
-            fontWeight: null,
-          },
-          lastActivityAt: new Date().toISOString(),
-        };
-
-        const epubIDBData: EpubIDBData = {
-          id: generatedEpubId,
-          kiwiId: generatedKiwiId,
-          file: bookInfo.file,
-          locations: bookInfo.locations,
-        };
-
-        await idb.add(IDBStore.KiwiStore, kiwiIDBData);
-        await idb.add(IDBStore.EpubStore, epubIDBData);
-        await idb.add(IDBStore.ParticipantStore, participantIDBData);
-
+        // if (abortControllerRef.current?.signal.aborted) {
+        //   return;
+        // }
+        const kiwiIDBData = await createKiwi(newKiwi);
         await router.invalidate();
-        setShareCode(generatedShareCode);
+        setShareCode(kiwiIDBData.shareCode);
         setStep(Step.Four);
       } catch (error) {
         setIsFailed(true);
@@ -138,11 +63,11 @@ function StepThree() {
   return isFailed ? (
     <FailedKiwi onGoBack={() => setStep(Step.Two)} />
   ) : (
-    <LoadingKiwi onCancel={handleCancel} />
+    <LoadingKiwi />
   );
 }
 
-function LoadingKiwi({ onCancel }: { onCancel: () => void }) {
+function LoadingKiwi() {
   return (
     <>
       <div className="flex flex-col items-center justify-center space-y-6 py-10">
@@ -166,11 +91,11 @@ function LoadingKiwi({ onCancel }: { onCancel: () => void }) {
           </p>
         </div>
       </div>
-      <DialogFooter className="sm:justify-between">
+      {/* <DialogFooter className="sm:justify-between">
         <Button onClick={onCancel} className="ml-auto">
           취소
         </Button>
-      </DialogFooter>
+      </DialogFooter> */}
     </>
   );
 }
